separate file for WinGLCanvas' events.
